+++ extensions/irc/test-support.ts
